rename saveContentButton to favouriteButton
