+++ components/Modules/VsBrFavouritesDisplay.vue
@@ -174,9 +174,9 @@ interface FavouriteCard {
 const fetchRequestStatus = ref('pending');
 const cardData = ref<FavouriteCard[]>([]);
 
-const favouritesEndpoint = configStore.featureFavouritesEndpoint;
+// const favouritesEndpoint = configStore.featureFavouritesEndpoint;
 // This will be removed before release
-// const favouritesEndpoint = 'https://release-brc.visitscotland.com/api/favourites/get-favourites';
+const favouritesEndpoint = 'https://release-brc.visitscotland.com/api/favourites/get-favourites';
 
 // Fetch CMS data for a list of UUIDs
 async function getSavedContentData(endpoint, data) {
